requirements.txt
django==3.2.5
django-bootstrap==0.2.4
whitenoise==5.2.0
gunicorn==20.1.0
pillow==8.3.1
django-taggit==1.5.0
markdown==3.3.4
